+++ frontend/components/configurations/tsmLinking/TsmLinkingOverviewTable.vue
@@ -15,41 +15,207 @@ SPDX-License-Identifier: EUPL-1.2
       sort-by="device"
       class="elevation-1"
       show-expand
+      :footer-props="{
+        itemsPerPageOptions:[50,100,-1]
+      }"
+      @click:row="toggleExpansion"
     >
       <template #[`item.deviceName`]="{ item }">
-        <ExtendedItemName :value="item.device" />
-      </template>
-      <template #[`item.licenseName`]="{ item }">
-        {{ item.licenseName }}
-        <a v-if="item.licenseUri" target="_blank" :href="item.licenseUri">
-          <v-icon small>
+        <DeviceShortNameSerialNumber
+          style="display:inline"
+          :value="item.device"
+        />
+        <v-btn
+          small
+          text
+          @click.stop.prevent="openDeviceLink(item.device)"
+        >
+          <v-icon
+            small
+          >
             mdi-open-in-new
           </v-icon>
-        </a>
+        </v-btn>
       </template>
-      <template #[`item.involvedDevices`]="{ item }">
-        {{ item.involvedDevices | sparseJoin }}
+      <template #[`item.measuredQuantity`]="{ item }">
+        <div v-if="item.deviceProperty">
+          <span v-if="item.deviceProperty.propertyName">
+            {{ item.deviceProperty.propertyName }}
+          </span>
+          <span
+            v-if="item.deviceProperty.label"
+            class="text--disabled"
+          >
+            - {{ item.deviceProperty.label }}
+          </span>
+          <span
+            v-if="item.deviceProperty.unitName"
+            class="text--secondary text-decoration-underline"
+          >
+            ({{ item.deviceProperty.unitName }})
+          </span>
+          <v-btn
+            small
+            text
+            @click.stop.prevent="openDevicePropertyLink(item.deviceProperty)"
+          >
+            <v-icon
+              small
+            >
+              mdi-open-in-new
+            </v-icon>
+          </v-btn>
+        </div>
+      </template>
+      <template #[`item.tsmdlEntityNames`]="{ item }">
+        <div>
+          <v-tooltip bottom>
+            <template #activator="{ on, attrs }">
+              <span
+                v-bind="attrs"
+                v-on="on"
+              >
+                {{ item.tsmEndpoint?.name ?? '-' }}
+              </span>
+            </template>
+            <span>TSM</span>
+          </v-tooltip>
+        </div>
+        <div>
+          <v-tooltip bottom>
+            <template #activator="{ on, attrs }">
+              <span
+                v-bind="attrs"
+                v-on="on"
+              >
+                {{ getDataSourceName(item) }}
+              </span>
+            </template>
+            <span>Datasource: {{ item.datasource?.name ?? '-' }}</span>
+          </v-tooltip>
+        </div>
+        <div>
+          <v-tooltip bottom>
+            <template #activator="{ on, attrs }">
+              <span
+                v-bind="attrs"
+                v-on="on"
+              >
+                {{ getThingName(item) }}
+              </span>
+            </template>
+            <span>Thing: {{ item.thing?.name ?? '-' }}</span>
+          </v-tooltip>
+        </div>
+        <div>
+          <v-tooltip bottom>
+            <template #activator="{ on, attrs }">
+              <span
+                v-bind="attrs"
+                v-on="on"
+              >
+                {{ getDatastreamName(item) }}
+              </span>
+            </template>
+            <span>Datastream: {{ item.datastream?.name ?? '-' }}</span>
+          </v-tooltip>
+        </div>
       </template>
       <template #[`item.actions`]="{ item }">
-        <v-icon
-          v-if="$auth.loggedIn && canModifyEntity(configuration)"
-          small
-          class="mr-2"
-          @click="editItem(item)"
-        >
-          mdi-pencil
-        </v-icon>
-        <v-icon
-          v-if="$auth.loggedIn && canModifyEntity(configuration)"
-          small
-          color="error"
-          @click="initDeleteDialog(item)"
-        >
-          mdi-delete
-        </v-icon>
+        <DotMenu>
+          <template #actions>
+            <DotMenuActionEdit
+              :readonly="!$auth.loggedIn || !canModifyEntity(configuration)"
+              @click="editItem(item)"
+            />
+            <DotMenuActionDelete
+              :readonly="!$auth.loggedIn || !canModifyEntity(configuration)"
+              @click="initDeleteDialog(item)"
+            />
+          </template>
+        </DotMenu>
       </template>
       <template #expanded-item="{ headers, item }">
         <td :colspan="headers.length">
+          <div class="ml-4">
+            <v-row class="mt-2">
+              <v-col
+                cols="4"
+                xs="4"
+                sm="3"
+                md="2"
+                lg="2"
+                xl="1"
+                class="font-weight-medium"
+              >
+                License
+              </v-col>
+              <v-col
+                cols="8"
+                xs="8"
+                sm="9"
+                md="4"
+                lg="4"
+                xl="5"
+                class="nowrap-truncate"
+              >
+                {{ item.licenseName }}
+                <a v-if="item.licenseUri" target="_blank" :href="item.licenseUri">
+                  <v-icon small>
+                    mdi-open-in-new
+                  </v-icon>
+                </a>
+              </v-col>
+            </v-row>
+            <v-row>
+              <v-col
+                cols="4"
+                xs="4"
+                sm="3"
+                md="2"
+                lg="2"
+                xl="1"
+                class="font-weight-medium"
+              >
+                Involved Devices
+              </v-col>
+              <v-col
+                cols="8"
+                xs="8"
+                sm="9"
+                md="4"
+                lg="4"
+                xl="5"
+                class="nowrap-truncate"
+              >
+                {{ item.involvedDevices | sparseJoin }}
+              </v-col>
+            </v-row>
+            <v-row>
+              <v-col
+                cols="4"
+                xs="4"
+                sm="3"
+                md="2"
+                lg="2"
+                xl="1"
+                class="font-weight-medium"
+              >
+                Aggregation
+              </v-col>
+              <v-col
+                cols="8"
+                xs="8"
+                sm="9"
+                md="4"
+                lg="4"
+                xl="5"
+                class="nowrap-truncate"
+              >
+                {{ item.aggregationText }}
+              </v-col>
+            </v-row>
+          </div>
           <tsm-linking-basic-data-table
             :linking="item"
           />
@@ -77,22 +243,55 @@ import { TsmLinking } from '@/models/TsmLinking'
 import { DeviceProperty } from '@/models/DeviceProperty'
 import { Device } from '@/models/Device'
 import {
-  DeleteConfigurationTsmLinkingAction,
-  ITsmLinkingState,
+  DeleteConfigurationTsmLinkingAction, FilteredLinkingsGetter,
   LoadConfigurationTsmLinkingsAction
 } from '@/store/tsmLinking'
 import { ConfigurationsState } from '@/store/configurations'
 import { CanModifyEntityGetter } from '@/store/permissions'
 import { SetLoadingAction } from '@/store/progressindicator'
+import DotMenu from '@/components/DotMenu.vue'
+import DotMenuActionEdit from '@/components/DotMenuActionEdit.vue'
+import DotMenuActionDelete from '@/components/DotMenuActionDelete.vue'
+import { shortenMiddle } from '@/utils/stringHelpers'
+import { dateToDateString } from '@/utils/dateHelper'
+import { TsmdlDatasource } from '@/models/TsmdlDatasource'
+import { TsmdlDatastream } from '@/models/TsmdlDatastream'
+import { TsmdlThing } from '@/models/TsmdlThing'
+import { TsmEndpoint } from '@/models/TsmEndpoint'
+import DeviceShortNameSerialNumber from '@/components/devices/DeviceShortNameSerialNumber.vue'
+
+type DataTableLinkingsItem = {
+  id: string
+  tsmdlEntityNames: string
+  offsets: string
+  startDate: string
+  endDate: string
+  measuredQuantity: string
+  device: Device | null
+  deviceName: string
+  deviceProperty: DeviceProperty | null
+  datasource: TsmdlDatasource | null
+  datastream: TsmdlDatastream | null
+  thing: TsmdlThing | null
+  tsmEndpoint: TsmEndpoint | null
+  licenseName: string
+  licenseUri: string
+  aggregationText: string
+  involvedDevices: Device[]
+}
 
 @Component({
   components: {
+    DeviceShortNameSerialNumber,
+    DotMenuActionDelete,
+    DotMenuActionEdit,
+    DotMenu,
     DeleteDialog,
     ExtendedItemName,
     TsmLinkingBasicDataTable
   },
   computed: {
-    ...mapState('tsmLinking', ['linkings']),
+    ...mapGetters('tsmLinking', ['filteredLinkings']),
     ...mapState('configurations', ['configuration']),
     ...mapGetters('permissions', ['canModifyEntity'])
   },
@@ -110,13 +309,10 @@ export default class TsmLinkingOverviewTable extends Vue {
 
   private headers = [
     {
-      text: 'TSM::Datasource::Thing::Datastream',
-      value: 'tsmdlEntityNames'
-    },
-    {
       text: 'Device',
       value: 'deviceName'
-    }, {
+    },
+    {
       text: 'Measured Quantity',
       value: 'measuredQuantity'
     }, {
@@ -128,16 +324,12 @@ export default class TsmLinkingOverviewTable extends Vue {
     }, {
       text: 'End date',
       value: 'endDate'
-    }, {
-      text: 'License',
-      value: 'licenseName'
-    }, {
-      text: 'Aggregation',
-      value: 'aggregationText'
-    }, {
-      text: 'Involved devices',
-      value: 'involvedDevices'
-    }, {
+    },
+    {
+      text: 'TSM/Datasource/Thing/Datastream',
+      value: 'tsmdlEntityNames'
+    },
+    {
       text: 'Actions',
       value: 'actions',
       sortable: false
@@ -148,28 +340,33 @@ export default class TsmLinkingOverviewTable extends Vue {
   private showDeleteDialog: boolean = false
   // vuex definition for typescript check
   configuration!: ConfigurationsState['configuration']
-  linkings!: ITsmLinkingState['linkings']
+  filteredLinkings!: FilteredLinkingsGetter
   deleteConfigurationTsmLinking!: DeleteConfigurationTsmLinkingAction
   loadConfigurationTsmLinkings!: LoadConfigurationTsmLinkingsAction
   canModifyEntity!: CanModifyEntityGetter
   setLoading!: SetLoadingAction
 
+  toggleExpansion (_item: TsmLinking, _row: { expand: (value: boolean) => void, isExpanded: boolean }) {
+    const { expand, isExpanded } = _row
+    expand(!isExpanded)
+  }
+
   get configurationId (): string {
     return this.$route.params.configurationId
   }
 
-  get dataTableLinkings (): Array<any> {
-    const dataTableLinkings: Array<any> = this.linkings
-    return dataTableLinkings.map((linking: TsmLinking) => {
+  get dataTableLinkings (): Array<DataTableLinkingsItem> {
+    return this.filteredLinkings.map((linking: TsmLinking) => {
       return {
         id: linking.id,
         tsmdlEntityNames: this.generateTsmdlEntityNames(linking),
         offsets: `(${linking?.deviceMountAction?.offsetX} | ${linking?.deviceMountAction?.offsetY} | ${linking?.deviceMountAction?.offsetZ})`,
-        startDate: this.$root.$options.filters!.toUtcDateTimeStringHHMM(linking.startDate),
-        endDate: this.$root.$options.filters!.toUtcDateTimeStringHHMM(linking.endDate),
+        startDate: dateToDateString(linking.startDate),
+        endDate: dateToDateString(linking.endDate),
         measuredQuantity: this.generatePropertyTitle(linking.deviceProperty),
         device: linking.device,
         deviceName: this.getDeviceName(linking.device),
+        deviceProperty: linking.deviceProperty,
         datasource: linking.datasource,
         datastream: linking.datastream,
         thing: linking.thing,
@@ -180,6 +377,16 @@ export default class TsmLinkingOverviewTable extends Vue {
         involvedDevices: linking.filterInvolvedDevices(this.devices)
       }
     })
+  }
+
+  get shortMiddleNumberBasedOnBreakpoint () {
+    switch (this.$vuetify.breakpoint.name) {
+      case 'xs': return 5
+      case 'sm': return 10
+      case 'md': return 15
+      case 'lg': return 25
+      case 'xl': return 40
+    }
   }
 
   editItem (item: TsmLinking) {
@@ -233,7 +440,50 @@ export default class TsmLinkingOverviewTable extends Vue {
     if (!value) {
       return ''
     }
-    return `${value.shortName} - ${value.manufacturerName} - ${value.model} - ${value.serialNumber}`
+
+    let nameString = value.shortName
+
+    if (value.serialNumber) {
+      nameString += ` (${value.serialNumber})`
+    }
+
+    return nameString
+  }
+
+  getDataSourceName (item: DataTableLinkingsItem) {
+    if (item.datasource?.name) {
+      return this.shortenbyAmount(item.datasource.name)
+    } else {
+      return '-'
+    }
+  }
+
+  getThingName (item: DataTableLinkingsItem) {
+    if (item.thing?.name) {
+      return this.shortenbyAmount(item.thing.name)
+    } else {
+      return '-'
+    }
+  }
+
+  getDatastreamName (item: DataTableLinkingsItem) {
+    if (item.datastream?.name) {
+      return this.shortenbyAmount(item.datastream.name)
+    } else {
+      return '-'
+    }
+  }
+
+  private shortenbyAmount (phrase: string): string {
+    return shortenMiddle(phrase, this.shortMiddleNumberBasedOnBreakpoint)
+  }
+
+  private openDeviceLink (device: Device) {
+    window.open(this.$router.resolve(`/devices/${device.id}`).href, '_blank')
+  }
+
+  private openDevicePropertyLink (deviceProperty: DeviceProperty) {
+    window.open(this.$router.resolve(`/measuredquantities/${deviceProperty.id}`).href, '_blank')
   }
 }
 </script>